Add functionality to page
Added initializeGrid, resetGrid, getGridSize, and
initializeGridSizeButton functions.

initializeGrid adds a mouseover event listener for each child appended
to the grid container that turns the tile black. Called when user
enters a grid size number.

resetGrid removes the children appended to the grid container. This is
called when the user enters a new size so they can start with a clean
slate.

getGridSize gets the desired grid size from the user.

initializeGridSizeButton adds a click event listener to the 'Grid Size'
button. This button calls getGridSize, initializeGrid, and resetGrid
functions.